backend/requirements.txt
requests>=2.28.0
pandas>=1.5.0
geopandas>=0.12.0
click>=8.1.0
python-dotenv>=1.0.0
tqdm>=4.64.0
folium>=0.14.0
sqlalchemy>=1.4.0
geopy>=2.3.0
shapely>=2.0.0
pyproj>=3.4.0
flask>=2.3.0
flask-cors>=4.0.0
plotly>=5.15.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0